+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to springcm-sdk will be documented in this file.
 
 ## [Unreleased]
+
+## [0.3.4] - 2019-12-11
 ### Added
 * Folder#delete
 * Document#delete
@@ -67,7 +69,7 @@ All notable changes to springcm-sdk will be documented in this file.
 ### Added
 * Initial release to reserve gem name
 
-[Unreleased]: https://github.com/paulholden2/springcm-sdk/compare/0.3.3...HEAD
+[Unreleased]: https://github.com/paulholden2/springcm-sdk/compare/0.3.4...HEAD
 [0.1.0]: https://github.com/paulholden2/springcm-sdk/releases/tag/0.1.0
 [0.1.1]: https://github.com/paulholden2/springcm-sdk/releases/tag/0.1.1
 [0.1.2]: https://github.com/paulholden2/springcm-sdk/releases/tag/0.1.2
@@ -76,3 +78,4 @@ All notable changes to springcm-sdk will be documented in this file.
 [0.3.1]: https://github.com/paulholden2/springcm-sdk/releases/tag/0.3.1
 [0.3.2]: https://github.com/paulholden2/springcm-sdk/releases/tag/0.3.2
 [0.3.3]: https://github.com/paulholden2/springcm-sdk/releases/tag/0.3.3
+[0.3.4]: https://github.com/paulholden2/springcm-sdk/releases/tag/0.3.4